DBT/project/requirements.txt
kafka-python 
PyMySQL
